requirements.txt
telethon
python-dotenv
SQLAlchemy